// src/name-generator/names/types.ts
export enum Race {
  Hyur = 'hyur',
  Elezen = 'elezen',
  Lalafell = 'lalafell',
  Miqote = 'miqote',
  Roegadyn = 'roegadyn',
  AuRa = 'auRa',
  Hrothgar = 'hrothgar',
  Viera = 'viera'
}

export enum Clan {
  Midlander = 'midlander',
  Highlander = 'highlander',
  Wildwood = 'wildwood',
  Duskwight = 'duskwight',
  Plainsfolk = 'plainsfolk',
  Dunesfolk = 'dunesfolk',
  SeekersOfTheSun = 'seekersOfTheSun',
  KeepersOfTheMoon = 'keepersOfTheMoon',
  SeaWolves = 'seaWolves',
  Hellsguard = 'hellsguard',
  Raen = 'raen',
  Xaela = 'xaela',
  Helions = 'helions',
  TheLost = 'theLost',
  Rava = 'rava',
  Veena = 'veena'
}

export enum Gender {
  Male = 'male',
  Female = 'female'
}